suggesed changes to make package more pythonic
