+++ adaptfx/__init__.py
@@ -5,3 +5,8 @@ Created on Thu Jul 22 10:58:43 2021
 @author: yoelh
 """
 
+import adaptfx_interpolation_2D
+import adaptfx_interpolation_3D
+
+import Interface_2D
+import Interface_3D
