--- adaptfx/discrete_programs/testfile
@@ -1,1 +1,0 @@
-test
