+++ pyproject.toml
@@ -1,6 +1,3 @@
 [build-system]
-requires = ["tkinter","numpy","scipy","scipy.stats","pandas",
-    "setuptools>=42",
-    "wheel"
-]
+requires = ["setuptools >= 42", "wheel"]
 build-backend = "setuptools.build_meta"
