+++ setup.cfg
@@ -1,24 +1,25 @@
 [metadata]
-name = example-pkg-YOEL
+name = adaptfx
 version = 0.0.1
 author = Yoel Pérez Haas
 author_email = yoel.perezhaas@uzh.ch
-description = A small example package
+description = Python package for computing adaptive fractionation plans
 long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/YoelPH/Adaptive-fractionation-2D
-project_urls =
-    Bug Tracker = 
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
 
 [options]
-package_dir =
-    = adaptfx
 packages = find:
+setup_requires = setuptools
 python_requires = >=3.6
+install_requires = 
+    numpy
+    scipy
+    pandas
 
 [options.packages.find]
 where = adaptfx
